credit-coop-system/member-portal/server/requirements.txt
psycopg2-binary==2.9.9
Pillow==10.0.1
Werkzeug==2.3.7
Flask==2.3.3
